% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{addActivity}
\alias{addActivity}
\alias{addActivity<-}
\alias{addActivity<-.butilsAccount}
\title{Add a new entry to the account.}
\usage{
addActivity(object, ...) <- value

\method{addActivity}{butilsAccount}(object, involved = NULL,
  type = as.character(NA), date = as.Date(NA), note = "", label = NULL,
  ...) <- value
}
\arguments{
\item{object}{the account.}

\item{...}{ignored.}

\item{value}{a named vector describing name paid what.}

\item{involved}{who was involved in the activity?}

\item{type}{a character string describing the activity.}

\item{date}{the date at which the activity happen.}

\item{note}{additional text.}
}
\description{
Add a new entry to the account.
}
\examples{
#### Create an account ####
myAcc <- createAccount()

#### Ease input with short names ####
addNickName(myAcc) <- list("Brice Ozenne" = "Brice",
                           "Sebastian Holst" = c("Seb","Sebastian"))

#### Add activity paid by one person for others ####
addActivity(myAcc, involved = c("Brice","Sebastian"), type = "Court") <- c("Sebastian" = 200)

#### Add expense, i.e. activity paid by nobody ####
addActivity(myAcc, involved = c("Brice","Sebastian"), type = "Shuttlecock") <- 100

#### Add contribution from one member ####
addActivity(myAcc, type = "Shuttlecock") <- c("Brice" = 100)

####
myAcc
}
