% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{addMatch}
\alias{addMatch}
\alias{addMatch<-}
\alias{addMatch<-.butilsAccount}
\title{Add a new match to the account}
\usage{
addMatch(object, ...) <- value

\method{addMatch}{butilsAccount}(object, involved = NULL,
  date = as.Date(NA), shuttlecock = NULL, price.shuttlecock = 175/12,
  refund = NULL, note = "", print.balance = TRUE,
  replace.entry = FALSE, ...) <- value
}
\arguments{
\item{object}{the account.}

\item{...}{ignored.}

\item{value}{[numeric] a named vector describing name paid what.}

\item{involved}{[vector of characters] the names of the players.}

\item{date}{[date] the date at which the activity happen.}

\item{shuttlecock}{[interger > 0] the number of shuttlecock used.}

\item{price.shuttlecock}{[numeric] the price of a single shuttlecock.}

\item{refund}{[list] A named list indicating who has refunded who.}

\item{note}{[character] additional text.}

\item{print.balance}{[logical] should the balance for the match be displayed.}

\item{replace.entry}{[logical] erase any entry that have the same name.}
}
\description{
Add a new match to the account.
}
\examples{
#### Create an account ####
myAcc <- createAccount()

#### Ease input with short names ####
addNickName(myAcc) <- list("Brice Ozenne" = "Brice",
                           "Sebastian Holst" = c("Seb","Sebastian"),
                           "Thomas S" = "Thomas")

#### Add a single - no shuttlecock ####
addMatch(myAcc, date = "12-06-2017", involved = c("Brice","Sebastian")) <- c("Sebastian" = 200)
addMatch(myAcc, date = "12-06-2017", involved = c("Brice","Sebastian","Sebastian")) <- c("Sebastian" = 200)
addMatch(myAcc, date = "15-06-17", shuttlecock = 5, 
                involved = c("Brice","Sebastian")) <- c("Thomas" = 200)

addActivity(myAcc,
           label = "Shuttlecock-Brice",
           date = as.Date("18-01-13"),
           note = "shuttlecock (1 tube)",
           type = "Buy-Shuttlecock") <- c("Brice" = 170)

}
