% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.boot.R
\name{as.boot}
\alias{as.boot}
\alias{as.boot.bootReg}
\title{Convertion to boot object}
\usage{
as.boot(object, index)

\method{as.boot}{bootReg}(object, index = NULL)
}
\arguments{
\item{object}{a bootReg object}

\item{index}{[optional] index of the observation to keep.}
}
\description{
Convertion to boot object
}
\examples{
#### data  ####
n <- 1e2
set.seed(10)
df.data <- data.frame(Y = rnorm(n),
                     group = gl(3, 5, n, labels = c("Ctl","Trt","Neu")),
                     gender = gl(2, 5, n, labels = c("Female","Male"))[sample.int(n)]
                     )

#### lm ####
m.lm <- lm(Y ~ group*gender, data = df.data)
resBoot <- bootReg(m.lm, n.boot = 1e1)
as.boot(resBoot)

}
