% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalCI.R
\name{autoplot.conditionalCI}
\alias{autoplot.conditionalCI}
\title{Display Conditional Confidence Intervals}
\usage{
\method{autoplot}{conditionalCI}(object, value = TRUE,
  unconditional.CI = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{output of the function \code{conditionalCI}.}

\item{value}{[logical] should the identity line be displayed (in blue)?}

\item{unconditional.CI}{[logical] should the unconditional confidence intervals be displayed (in red)?}

\item{plot}{[logical] should the graph be displayed in a graphical window?}
}
\description{
Display conditional confidence intervals.
}
