% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoA.R
\name{autoplot.LoA}
\alias{autoplot.LoA}
\title{Bland and Altman Plot with Limits of Agreement}
\usage{
\method{autoplot}{LoA}(object, display.ci = TRUE, plot = TRUE,
  line.size = 1.5, alpha = 0.25, name.legend = "",
  labels = c("bias", "LoA (lower)", "LoA (upper)"), colors = c("red",
  "blue", "forestgreen"), linetype = c(1, 2, 2), ...)
}
\arguments{
\item{object}{An object of class \code{LoA}, output of the \code{LoA} function.}

\item{display.ci}{[logical] should the confidence intervals be displayed?}

\item{plot}{[logical] should the plot be printed?}

\item{line.size}{[numeric >0] the width of the horizontal lines.}

\item{alpha}{[numeric 0-1] transparency parameter for the confidence intervals.}

\item{name.legend}{[character] character string displayed in the caption.}

\item{labels}{[character of length 3] text shown in the caption for the bias, lower, and upper LoA.}

\item{colors}{[character of length 3] colors used to display the bias, lower, and upper LoA.}

\item{linetype}{[numeric of length 3] type of lines used to display the bias, lower, and upper LoA.}

\item{...}{not used, for compatibility with the generic method.}
}
\description{
Display a Bland and Altman plot with limits of agreement.
}
