% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPb.R
\name{bootPb}
\alias{bootPb}
\title{Copy of \code{boot::boot} with progress bar}
\usage{
bootPb(data, statistic, R, sim = "ordinary", stype = c("i", "f", "w"),
  strata = rep(1, n), L = NULL, m = 0, weights = NULL,
  ran.gen = function(d, p) d, mle = NULL, simple = FALSE, ...,
  parallel = c("no", "multicore", "snow"),
  ncpus = getOption("boot.ncpus", 1L), cl = NULL)
}
\arguments{
\item{data}{see \code{boot::boot}}

\item{statistic}{see \code{boot::boot}}

\item{R}{see \code{boot::boot}}

\item{sim}{see \code{boot::boot}}

\item{stype}{see \code{boot::boot}}

\item{strata}{see \code{boot::boot}}

\item{L}{see \code{boot::boot}}

\item{m}{see \code{boot::boot}}

\item{weights}{see \code{boot::boot}}

\item{ran.gen}{see \code{boot::boot}}

\item{mle}{see \code{boot::boot}}

\item{simple}{see \code{boot::boot}}

\item{...}{see \code{boot::boot}}

\item{parallel}{see \code{boot::boot}}

\item{ncpus}{see \code{boot::boot}}

\item{cl}{see \code{boot::boot}}
}
\value{
see \code{boot::boot}
}
\description{
Copy of \code{boot::boot} with progress bar
}
\details{
Copied from https://stackoverflow.com/questions/37673931/add-a-progress-bar-to-boot-function-in-r
}
\examples{

m.boot <- function(data, indices) {
  d <- data[indices]
  mean(d, na.rm = TRUE) 
}
tot_rep <- 200
v1 <- rnorm(1000)
b <- bootPb(v1, m.boot, R = tot_rep)
}
