% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPartialResiduals.R
\name{calcPartialResiduals}
\alias{calcPartialResiduals}
\title{Compute partial residuals for linear and linear mixed models.}
\usage{
calcPartialResiduals(model, var, keep.intercept = FALSE,
  conditional = FALSE, interval = "confidence", level = 0.95,
  npoints = 100, quantile.range = c(0, 1), FUN.df, ...)
}
\arguments{
\item{model}{Model object (e.g. \code{lm})}

\item{var}{[character] Independent variable(s) whose effect should be kept in the partial residuals.}

\item{keep.intercept}{[logical] should the partial residual be computed keeping the contribution of the reference level?}

\item{conditional}{[logical] are the predictions conditional to the random effect? (if any)}

\item{interval}{[character] Type of interval calculation ("confidence" or "prediction").}

\item{level}{[numeric, 0-1] Level of confidence limits (default 95\%)}

\item{npoints}{[integer] Length of the vector of unique values relative to each continuous variable.}

\item{quantile.range}{[numeric vector, 0-1] the quantiles (for the continous covariates) between which the fitted values will be computed.}

\item{FUN.df}{Optional function returning the residual degree of freedoms.}

\item{\dots}{additional arguments to lower level functions}
}
\value{
list with following members:
\item{data}{Original dataset with an additional column containing the partial residuals}
\item{partialFit}{Fitted values according to the range of values of var}
\item{var}{variable(s) for which the partial residuals have been computed}
\item{level}{Level of confidence}
\item{name.Y}{name of the response variable}
}
\description{
Compute partial residuals for linear and linear mixed models.
}
\details{
In a linear model:
\deqn{Y = \alpha + \beta X + \gamma Z + \varepsilon}
The partial residuals are defined by:
\deqn{\varepsilon_{X} = \beta X + \varepsilon}
or
\deqn{\varepsilon_{X} = \alpha + \beta X + \varepsilon}
depending on the value of the argument \code{keep.intercept}.
The X matrix is defined contains the variables defined by the \code{var} argument.

When using mixed models, the confidence and prediction intervals
ignore the uncertainty of the covariance parameters.
}
\examples{
library(lava)
set.seed(10)
m.lvm <- lvm(Y~X1+X2+Id+Id2)
categorical(m.lvm, K = 5) <- ~Id+Id2
d <- lava::sim(n = 1e2, m.lvm)
 
m <- lm(Y~X1+X2, data = d)
pres0 <- calcPartialResiduals(m, var = NULL)
range(residuals(m) - pres0$data$pResiduals)

pres1 <- calcPartialResiduals(m, var = "X1")
fit1 <- coef(m)["(Intercept)"] + coef(m)["X2"] * d$X2
range((d$Y - fit1) - pres1$data$pResiduals)
cor(d$X1, pres1$data$pResiduals)
lava::partialcor(~X2,d[,c("Y","X1","X2")])

pres2 <- calcPartialResiduals(m, var = c("X1","X2"))
fit2 <- coef(m)["(Intercept)"]
range((d$Y - fit2) - pres2$data$pResiduals)

}
\seealso{
\code{plot.partialResiduals} for a graphical display and more examples.'
}
\author{
Brice Ozenne
}
\keyword{regression}
