% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{calcR2}
\alias{calcR2}
\title{Compute the R square using the predicted values}
\usage{
calcR2(model)
}
\arguments{
\item{model}{the model from which the R squared should be computed}

\item{Y}{the observed response}

\item{model0}{the null model}
}
\description{
Compute the R square using the predicted values
}
\examples{
m <- lvm(Y~X1+X2+X3)
d <- sim(m, 1e2)

m.lm <- lm(Y~X1+X2+X3, data = d)
calcR2(m.lm)

summary(m.lm)$r.squared
}
