% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalCI.R
\name{confint.conditionalCI}
\alias{confint.conditionalCI}
\title{Conditional confidence interval}
\usage{
\method{confint}{conditionalCI}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{output of the function \code{conditionalCI}}

\item{parm}{not used. For compatibility with the generic method.}

\item{level}{not used. For compatibility with the generic method.}

\item{...}{not used. For compatibility with the generic method.}
}
\description{
Extract the conditional confidence interval from the object.
}
