% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_correlationMatrix.R
\name{cor.testDT}
\alias{cor.testDT}
\title{Perform pairwise correlation tests between matrix columns}
\usage{
cor.testDT(data, names = NULL, format, lower.tri = TRUE,
  method.cor = "cor.test", use.pairwiseNNA = TRUE, reorder = "AOE",
  imput.value = 0, hclus.method = "complete", trace = TRUE, plot = TRUE,
  args.plot = list(), output = "data.table", ...)
}

