% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{create_github_action_files}
\alias{create_github_action_files}
\title{Automatically create package files for GitHub Actions}
\usage{
create_github_action_files(quiet = TRUE)
}
\arguments{
\item{quiet}{Should all output be suppressed?}
}
\description{
Automatically create package files for GitHub Actions
}
