% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{create_package_files}
\alias{create_package_files}
\title{Automatically create package files}
\usage{
create_package_files(quiet = TRUE, lang = "en-GB", github_actions = TRUE)
}
\arguments{
\item{quiet}{Should all output be suppressed?}

\item{lang}{Preferred spelling language. Usually either \code{"en-US"} or
\code{"en-GB"}.}

\item{github_actions}{Should GitHub Action files be written?}
}
\description{
Automatically create package files
}
