% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{create_package_files_for_github_actions}
\alias{create_package_files_for_github_actions}
\title{Automatically create package files for GitHub Actions}
\usage{
create_package_files_for_github_actions(quiet = TRUE)
}
\arguments{
\item{quiet}{Should all output be suppressed?}
}
\description{
Automatically create package files for GitHub Actions
}
