% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{createAccount}
\alias{createAccount}
\title{Create an empty account}
\usage{
createAccount(nickName = NULL)
}
\arguments{
\item{nickName}{a named list to be passed to \code{addNickName<-}.}
}
\description{
Create an empty account.
}
