% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedCompMean.R
\name{.allPermutation}
\alias{.allPermutation}
\title{Generate all possible permutations}
\usage{
.allPermutation(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Generate all possible permutations
}
\examples{
.allPermutation(1)
.allPermutation(1:2)
.allPermutation(1:3)
}
\references{
https://stackoverflow.com/questions/11095992/generating-all-distinct-permutations-of-a-list-in-r
}
\keyword{internal}
