% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalInParentEnv.R
\name{evalInParentEnv}
\alias{evalInParentEnv}
\title{Find object in the parent environments}
\usage{
evalInParentEnv(name, envir)
}
\arguments{
\item{name}{character string containing the name of the object to get.}

\item{envir}{the environment from which to look for the object.}
}
\description{
Internal function
}
\examples{

e.lm <- lm(Y~X, data = data.frame(Y=1:5,X=1:5))
butils:::evalInParentEnv(e.lm$call$data, envir = environment())

}
