% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRchunk.R
\name{extractRchunk}
\alias{extractRchunk}
\title{Extract code chunks from an .org file}
\usage{
extractRchunk(
  file,
  newfile = NULL,
  file.header = NULL,
  file.start.SRC = NULL,
  file.end.SRC = NULL,
  rm.export.none = TRUE,
  rm.noexport = TRUE,
  index.chunk = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{file}{name of the file from which the R chunks should be extracted.}

\item{newfile}{name of the R file be created.}

\item{file.header}{string indicating headers.}

\item{file.start.SRC}{string indicating the start of a code block.}

\item{file.end.SRC}{string indicating the end of a code block.}

\item{rm.export.none}{should chunks of code with \code{:export none} be ignored?}

\item{rm.noexport}{should header with \code{:noexport:} be ignored?}

\item{index.chunk}{should the code block be numbered in the .R file?}

\item{overwrite}{Can the existing R file be overwritten?}
}
\description{
Collect code chunks from an .org file and put them into an .R file.
}
