% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSRCorgmode.R
\name{extractSRCorgmode}
\alias{extractSRCorgmode}
\title{Extract code chunks from an .org file}
\usage{
extractSRCorgmode(file, file.header = "\\\\*",
  file.start.SRC = "\\\\#\\\\+BEGIN_SRC R",
  file.end.SRC = "\\\\#\\\\+END_SRC", index.chunk = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{file}{name of the org file}

\item{file.header}{string indicating headers}

\item{file.start.SRC}{string indicating the start of a code block}

\item{file.end.SRC}{string indicating the end of a code block}

\item{index.chunk}{should the code block be numbered in the .R file?}

\item{overwrite}{Can the existing R file be overwritten?}
}
\description{
Collect code chunks from an .org file and put them into an .R file.
}
