% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialResiduals.R
\name{FUN.predict}
\alias{FUN.predict}
\alias{predict_merTools}
\alias{predict_AICcmodavg}
\title{prediction function for partialResiduals}
\usage{
predict_merTools(object, newdata, level, conditional, interval, ...)

predict_AICcmodavg(object, newdata, level, conditional, interval, ...)
}
\arguments{
\item{object}{a model}

\item{newdata}{the dataset used to make the predictions}

\item{level}{the confidence level for the intervals}

\item{conditional}{are the predictions conditional to the random effect? (if any)}

\item{interval}{the type of interval, prediction or confidence}

\item{...}{not used}
}
\description{
Suggestion of functions to obtain confidence intervals with mixed models
}
\details{
predict_merTools is based on the \code{predictInterval} function of the merTools package
and is only compatible with \code{lmer} models.

predict_AICcmodavg is based on the \code{predictSE} function of the AICcmodavg package
and is compatible with \code{lmer}, \code{lme}, and \code{gls} models. It can only be used to obtain confidence intervals.

See the details of the documentation of these functions for more information about the validity of the intervals.
}
