% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigmaGLS.R
\name{getSigmaGLS}
\alias{getSigmaGLS}
\title{Variance-covariance of GLS object}
\usage{
getSigmaGLS(gls, type = "covariance", upper = NULL, individual = NULL,
  addId = TRUE, plot = TRUE, args.plot = list(), output = "matrix",
  trace = 1)
}
\arguments{
\item{gls}{the gls object}

\item{type}{Can be either "correlation" or "covariance".}

\item{upper}{logical value indicating whether the upper triangle of the distance matrix should be returned}

\item{individual}{cluster for which the variance covariance matrix should be returned}

\item{addId}{the id name at the row and col names}

\item{plot}{should the correlation matrix be displayed. \emph{logical}.}

\item{args.plot}{a list of arguments to be passed to \code{ggHeatmap} to specify how the correlation matrix should be displayed}

\item{output}{how to output the correlation value. Can be \code{matrix}, \code{data.table} or \code{plot}.}

\item{trace}{should the progression of the computation of the correlation be displayed. \emph{logical}.}
}
\description{
Rebuilt the variance-covariance matrix from a GLS object
}
\examples{
if(require(nlme)){
data(Ovary)
Ovary <- as.data.table(Ovary)
setkeyv(Ovary, "Mare")
fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), 
            data = Ovary, correlation = corAR1(form = ~ 1 | Mare))
getSigmaGLS(fm1)
}
}
\keyword{correlation}
\keyword{display}
\keyword{function}
\keyword{gls}

