% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCor.R
\name{ggCor}
\alias{ggCor}
\title{Display correlation/covariance  matrix}
\usage{
ggCor(data, type = "correlation", legend_title = NULL, plot = TRUE,
  ...)
}
\arguments{
\item{data}{a matrix containing in each columns the observations for each variable.}

\item{type}{should the raw values, correlation or covariance be displayed?}

\item{legend_title}{argument to be passed to \code{\link{ggHeatmap}}.}

\item{plot}{should the graphic be plotted?}

\item{...}{additionnal parameters to be passed to \code{\link{ggHeatmap}}.}
}
\description{
Display correlation/covariance matrix
}
\examples{

n <- 100
X <- cbind(X1=rnorm(n),X2=rnorm(n),X3=rnorm(n))

ggCor(X)
}
