% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.testDT.R
\name{ggHeatmap}
\alias{ggHeatmap}
\title{Display pairwise correlation}
\usage{
ggHeatmap(data, name.x, name.y, name.fill, add.text, round = NULL,
  title = "", xlab = "", ylab = "", legend_title = "correlation",
  na.value = "grey50", col_low = "blue", col_midpoint = "white",
  col_high = "red", midpoint = 0, limits = NULL, textSize = 15,
  angle.x = 90)
}
\arguments{
\item{data}{the dataset containing the correlation values.}

\item{name.x}{name of the column containing the label of the first outcome involved in the correlation}

\item{name.y}{name of the column containing the label of the second outcome involved in the correlation}

\item{name.fill}{name of the column containing the value of the correlation}

\item{add.text}{additional information to be displayed above the squares representing the correlation. Must name a column in data.}

\item{round}{if not \code{NULL}, the number of digit used to round add.text using signif. Can also be \code{p.value} to indicates with the usual convention the significance level.}

\item{title}{the title of the plot}

\item{xlab}{the name of the x axis}

\item{ylab}{the name of the y axis}

\item{legend_title}{the name of the legend}

\item{na.value}{the color used to display missing values (NA).}

\item{col_low}{the color used to diplay low correlation values}

\item{col_midpoint}{the color used to diplay intermediate correlation values}

\item{col_high}{the color used to diplay high correlation values}

\item{midpoint}{the value corresponding to intermediate correlation value.}

\item{limits}{the minimum and maximum correlation values used to build the color panel}

\item{textSize}{the size of the text in the plot}

\item{angle.x}{the inclination of the x labels}
}
\value{
a ggplot object
}
\description{
Display a correlation matrix (or heatmap) using ggplot2
}
\details{
data must be coercible to data.table. \cr
}
\keyword{correlation}
\keyword{display}
\keyword{function}

