% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggResPlot.R
\name{ggResPlot}
\alias{ggResPlot}
\alias{ggResPlot.lvmfit}
\title{Residual plot using ggplot2}
\usage{
ggResPlot(object, x, plot, ...)

\method{ggResPlot}{lvmfit}(object, x, plot = TRUE, geom = "point", ...)
}
\arguments{
\item{object}{a \code{lvm} object.}

\item{x}{[character] the name of an observed variable.}

\item{plot}{[logical] should the plot be displayed?.}

\item{...}{arguments to be passed to lower level functions}

\item{geom}{[character] the type of plot to make:
can be \code{"point"} for a scatterplot or \code{"boxplot"} for boxplots.}
}
\description{
Display the residuals as a function of a variable
}
