% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSurv.R
\name{ggSurv}
\alias{ggSurv}
\alias{ggSurv.coxph}
\alias{ggSurv.data.table}
\alias{ggSurv.survfit}
\title{Survival curve using ggplot2}
\usage{
ggSurv(x, ...)

\method{ggSurv}{survfit}(x, ...)

\method{ggSurv}{coxph}(x, ...)

\method{ggSurv}{data.table}(x, format = "data.table", timeVar = "time",
  survivalVar = "survival", ciInfVar = NULL, ciSupVar = NULL,
  eventVar = NULL, censorVar = NULL, strataVar = NULL, plot = TRUE,
  legend.position = "top", title = NULL, textSize = NULL, ylim = NULL,
  line.size = 2, confint = FALSE, alpha.CIfill = 0.2,
  alpha.CIline = 0.5, censoring = FALSE, alpha.censoring = 1,
  colour.censoring = rgb(0.2, 0.2, 0.2), shape.censoring = 3,
  size.censoring = 2, name.censoring = "censoring", events = FALSE,
  alpha.events = 1, colour.events = rgb(0.2, 0.2, 0.2), shape.events = 8,
  size.events = 2, name.events = "event", ...)
}
\arguments{
\item{x}{a coxph object or a survfit object or a data.table object.}

\item{...}{arguments to be passed to lower level functions}

\item{format}{the format used to export the data. Can be data.table or data.frame.}

\item{timeVar}{the name of the column in the data.table object containing the time variable}

\item{survivalVar}{the name of the column in the data.table object containing the survival at each time}

\item{ciInfVar}{the name of the column in the data.table object containing the lower bound of the confidence interval}

\item{ciSupVar}{the name of the column in the data.table object containing the upper bound of the confidence interval}

\item{eventVar}{the name of the column in the data.table object containing the number of events at each time}

\item{censorVar}{the name of the column in the data.table object containing the number of censored patients at each time}

\item{strataVar}{the name of the column in the data.table object containing the strata variable}

\item{plot}{should the plot be displayed}

\item{legend.position, title, textSize}{annotations on the plot}

\item{ylim}{y range (survival)}

\item{line.size}{size of the survival curves}

\item{confint, alpha.CIfill, alpha.CIline}{how to display confidence intervals}

\item{censoring, alpha.censoring, colour.censoring, shape.censoring, size.censoring, name.censoring}{how to display censored events}

\item{events, alpha.events, colour.events, shape.events, size.events, name.events}{how to display non-censored events}
}
\description{
Display a non-parametric or semi-parametric survival curve. Can handle one strata variable.
}
\examples{
library(survival)
dt <- as.data.table(aml)
dt[,x:=as.factor(x)]

## survfit
KM <- survfit(Surv(time, status) ~ x, data = dt)
ggSurv(KM, confint = FALSE)

## coxph
Cox <- coxph(Surv(time, status) ~ x, data = dt, x = TRUE)
ggSurv(Cox)

## data.table
dt2 <- data.table(time = 1:10, 
                  survival = seq(1, by = -0.01, length.out = 10), 
                  n.censor = 0, n.event = 1)
ggSurv(dt2)

}

