% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_DetectOutliers.R
\name{identifyOutlier}
\alias{factorOutlier}
\alias{identifyOutlier}
\alias{numOutlier}
\title{Identify outlier}
\usage{
numOutlier(x, type = "auto", th.gaussian = 3, th.hampel = 3,
  th.boxplot = 1.5, na.rm = FALSE)

factorOutlier(x, threshold = 0.01, useNA = "ifany")
}
\arguments{
\item{x}{a vector of numeric}

\item{type}{the type of robust metric for assessing the "average" value}

\item{th.gaussian}{the threshold for defining an outlier when using the mean}

\item{th.hampel}{the threshold for defining an outlier when mad}

\item{th.boxplot}{the threshold for defining the IQR}

\item{na.rm}{should na be removed.}

\item{threshold}{the threshold for defining an outlier regarding the prevalence (categorical variables)}

\item{useNA}{argument passed to \code{table}}
}
\description{
Identify numeric or factor outliers
}
\examples{
\dontrun{
numOutlier(rnorm(1e3))
}
}
\references{
idea from http://www.r-bloggers.com/finding-outliers-in-numerical-data/
}

