% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_SourceGitHub.R
\name{install_githubCommit}
\alias{install_githubCommit}
\title{Install commit}
\usage{
install_githubCommit(repo, username = NULL, ref = "master", subdir = NULL,
  host = "https://api.github.com", temporary = TRUE, force = FALSE,
  quiet = FALSE, ...)
}
\arguments{
\item{repo}{Repository address in the format username/repo}
}
\description{
Install a specific commit of a package, possibly under a temporary name
}
\details{

}
\examples{
install_githubCommit("bozenne/BuyseTest", ref = "70e5ef6c3c66d44f9227297225e2b9d2fe6acc06")


}
\keyword{function}
\keyword{github}
\keyword{package}

