% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_BC.R
\name{lmBC}
\alias{lmBC}
\title{Box Cox transformation for linear models}
\usage{
lmBC(formula, data, trace = TRUE, ...)
}
\arguments{
\item{formula}{same as lm}

\item{data}{same as lm}

\item{trace}{should the execution of the function be traced}

\item{...}{other arguments to be passed to lm}
}
\description{
Ouptut the linear model fitted on the Box-Cox transformed outcome
}
\examples{
if(require(lava)){
df <- sim(lvm(Y ~ X1 + X2),1e2)
df$Y <- df$Y-min(df$Y)+1
lm(Y~X1+X2, data = df)
lmBC(Y~X1+X2, data = df)
#' 
}

}
