% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoA.R
\name{LoA}
\alias{LoA}
\title{Limits of Agreement}
\usage{
LoA(x, y, conf.level = 0.95, LoA.multiplier = qnorm(0.975))
}
\arguments{
\item{x}{[numeric vector] vector containing the first measurement, each value corresponding to an individual.}

\item{y}{[numeric vector] vector containing the second measurement, each value corresponding to an individual (same as in \code{x}).}

\item{conf.level}{[numeric 0-1] confidence level of the confidence intervals.}

\item{LoA.multiplier}{[numeric 0-1] confidence level for the limits of agreements.}
}
\description{
Computes the limits of agreement between two measurements.
}
\details{
The bias is estimated as the average difference between y and x, i.e. bias = E[y-x].
The upper/lower limit of the limit of agreement are estimated as the bias plus/minus the LoA.multiplier times the standard deviation of y-x,
i.e. E[y-x] +/- LoA.multiplier * sqrt(Var[y-x]).

The standard errors, confidence intervals, and p-values are computed for the limit of the limit of agreement are computed using a delta method.
}
\examples{

set.seed(10)
X <- rnorm(100)
Y <- X + rnorm(100)

e.loa <- LoA(x = X, y = Y)
e.loa
autoplot(e.loa)

if(require(blandr)){ ## same results as blandr.statistics
 e.blandr <- blandr.statistics(method1 = Y, method2 = X)
 e.loa$LoA$estimate - c(e.blandr$bias, e.blandr$lowerLOA, e.blandr$upperLOA)
 e.loa$LoA$lower - c(e.blandr$biasLowerCI, e.blandr$lowerLOA_lowerCI, e.blandr$upperLOA_lowerCI)
 e.loa$LoA$upper - c(e.blandr$biasUpperCI, e.blandr$lowerLOA_upperCI, e.blandr$upperLOA_upperCI)
}
}
