% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massTtestCI.R
\name{massTtestCI}
\alias{massTtestCI}
\title{Mass t-tests with conditional inference}
\usage{
massTtestCI(X, Y, threshold, conf.level = 0.95,
  method.p.adjust = "bonferroni", ...)
}
\arguments{
\item{X}{[numeric matrix] the design matrix for the individuals of the control group.}

\item{Y}{[numeric matrix] the design matrix for the individuals of the experimental group.}

\item{threshold}{[numeric] the selection threshold for the difference in mean between the groups.}

\item{conf.level}{[numeric, 0-1] Confidence level.}

\item{method.p.adjust}{[character] Method used to adjust for multiple comparisons.
See argument method in \code{p.adjust} for more details.}

\item{...}{additional arguments passed to the function \code{conditionalCI}.}
}
\value{
A data.frame where each site corresponds to a row and with the following columns: \itemize{
\item site: the index of the site
\item estimate: the estimated mean difference
\item sigma: the estimated standard error of the estimated mean difference
\item statistic: the value of the test statistic
\item p.value: the (unadjusted) p-value 
\item adjusted.p.value: the p-value adjusted for multiple comparisons over all sites
\item select: whether the site has been selected (i.e. signal above the threshold)
\item lower.CI: lower bound of the post-selection confidence interval (unadjusted)
\item upper.CI: upper bound of the post-selection confidence interval (unadjusted)
\item adjusted.lower.CI: lower bound of the post-selection confidence interval (adjusted for multiple comparisons over all selected sites)
\item adjusted.upper.CI: upper bound of the post-selection confidence interval (adjusted for multiple comparisons over all selected sites)
}
}
\description{
Perform t-tests for all variables or only for variables showing a sufficiently large effect.
In the latter case, conditional confidence intervals are used instead of the "normal" confidence intervals
to account for the selection step.
}
\examples{

if(require(sp) & require(gstat)){

## settings
n.samples <- 20
n.sites <- 30*30

coords <- expand.grid(x = 1:sqrt(n.sites),
                     y = 1:sqrt(n.sites))
coords$mu <- 0
coords$mu[coords$x<=5 & coords$y<=5] <- 1

vario <- vgm(psill = 1, range = 2, model='Exp')

## simulate data
set.seed(10)

X <- sim2Dimage(n = n.samples,
                coords = coords[,c("x","y")],
                mu = 0,
                vgm = vario)$X

Y <- sim2Dimage(n = n.samples,
                coords = coords[,c("x","y")],
                mu = coords$mu,
                vgm = vario)$X

coords$meanX <- colMeans(X)
coords$meanY <- colMeans(Y)

spdf <- coords
spdf$effect <- coords$meanY - coords$meanX
gridded(spdf) <- ~x+y
effect.range <- range(spdf[["effect"]])

plot(spdf[,"effect"], zlim = effect.range)

res <- massTtestCI(X, Y, threshold = 0.5, method.CI = "shortest2")

## traditional analysis
spdf$p.value <- res$p.value
spdf$adjusted.p.value <- res$adjusted.p.value

plot(spdf[,"effect"], zlim = effect.range,
main = "traditional t-test")
points(spdf[spdf$adjusted.p.value<0.05,"p.value"], col = "green", lwd = 3)

## post selection CI
spdf$selected <- res$selected
spdf$lower.CI <- res$lower.CI
spdf$adjusted.lower.CI <- res$adjusted.lower.CI

plot(spdf[spdf$selected>0,"lower.CI"], zlim = effect.range,
     main = "post-selection unadjusted")
points(spdf[which(spdf$lower.CI>0),"lower.CI"], col = "green", lwd = 3)

plot(spdf[spdf$selected>0,"adjusted.lower.CI"], zlim = effect.range,
     main = "post-selection adjusted")
points(spdf[which(spdf$adjusted.lower.CI>0),"lower.CI"], col = "green", lwd = 3)
}

}
