% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{matrixTriangle}
\alias{matrixTriangle}
\alias{upperTriangle}
\alias{upperTriangle<-}
\alias{lowerTriangle}
\alias{lowerTriangle<-}
\title{Extract or replace the upper/lower triangular portion of a matrix.}
\usage{
upperTriangle(x, diag = FALSE, byrow = FALSE)

upperTriangle(x, diag = FALSE, byrow = FALSE) <- value

lowerTriangle(x, diag = FALSE, byrow = FALSE)

lowerTriangle(x, diag = FALSE, byrow = FALSE) <- value
}
\arguments{
\item{x}{[matrix]}

\item{diag}{[logical]
If \code{TRUE}, include the matrix diagonal.}

\item{byrow}{[logical]
If \code{FALSE}, return/replace elements in column-wise order.
If \code{TRUE}, return/replace elements in row-wise order.}

\item{value}{Either a single value or a vector of length equal to that of
the current upper/lower triangular.  Should be of a mode
which can be coerced to that of 'x'.}
}
\description{
Extract or replace the upper/lower triangular portion of a matrix
}
\details{
Essentially a copy from the R package gdata,
to limit the number of dependency in butils.
}
\examples{
## see the documentation of gdata::upperTriangle
}
