% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskIDM.R
\name{model.tables.riskIDM}
\alias{model.tables.riskIDM}
\title{Extract Probabilities From IDM}
\usage{
\method{model.tables}{riskIDM}(
  x,
  time = "unique",
  indiv = FALSE,
  state = utils::tail(x$states$name, 1),
  contrast = NULL,
  metric = "difference",
  ...
)
}
\arguments{
\item{time}{[numeric vector] time at which the probabilities should be extracted. Can be \code{"unique"} to extract at jump times (i.e. non-duplicated risk values).}

\item{indiv}{[logical] should covariate specific probabilities be extracted?}

\item{state}{[character] state relative to which the occupancy probabilities should be extracted.}

\item{contrast}{[character vector] optional argument indicating scenario that are to be compared.}

\item{metric}{[character] how to compare the probabilities between two scenarios: \code{"difference"} (alternative - reference) or \code{"ratio"} (alternative / reference).}

\item{object}{[riskIDM] output of the \code{riskIDM} function.}
}
\description{
Extract occupancy probabilities, difference, or ratio between probabilities from an illness death model.
}
