% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_normalize.R
\name{normalize}
\alias{normalize}
\title{Transform a variable to obtain approximate normality}
\usage{
normalize(X, na.rm = FALSE)
}
\arguments{
\item{X}{the vector of values}

\item{na.rm}{should the NA values be ignored when estimating the cumulative distribution function.}
}
\description{
(Slow) implementation of the transformation described in Albada, 2007.
}
\examples{

n <- 1000

## normal distribution ##
X <- rnorm(n)
Xnorm <- normalize(X)
shapiro.test(Xnorm)
# plot(X, Xnorm)

X.NA <- c(NA,X,NA)
Xnorm.NA <- normalize(X.NA, na.rm = TRUE)

## gamma distribution
X <- rgamma(n, shape = 1)
shapiro.test(X)
# hist(X)

Xnorm <- normalize(X)
shapiro.test(Xnorm)
# hist(Xnorm)
# plot(X,Xnorm)
}
\references{
Albada et al. Transformation of arbitrary distributions to the normal distribution with application to EEG test-retest reliability (2007, journal of Neuroscience Methods)
}

