% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{package_magic}
\alias{package_magic}
\title{Automatically create, update, and validate package files}
\usage{
package_magic(
  quiet = TRUE,
  github_actions = TRUE,
  rcmdcheck_args = "--no-manual",
  error_on = "note"
)
}
\arguments{
\item{quiet}{Should all output be suppressed?}

\item{github_actions}{Should GitHub Action files be written?}

\item{rcmdcheck_args}{Passed on to \code{args} argument in \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}.}

\item{error_on}{Whether to throw an error on \code{R CMD check} failures.
Note that the check is always completed (unless a timeout happens),
and the error is only thrown after completion. If \code{"never"}, then
no errors are thrown. If \code{"error"}, then only \code{ERROR} failures
generate errors. If \code{"warning"}, then \code{WARNING} failures generate
errors as well. If \code{"note"}, then any check failure generated an
error.}
}
\description{
Automatically create, update, and validate package files
}
\examples{
\dontrun{
package_magic()
}
}
