% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialCorrelation.R
\name{partialCorrelation}
\alias{partialCorrelation}
\alias{partialCorrelation.lm}
\alias{partialCorrelation.mmm}
\title{Compute the Partial Correlation Between the Outcome and a Covariate}
\usage{
partialCorrelation(object, var, fisher.transform, cluster)

\method{partialCorrelation}{lm}(object, var, fisher.transform = FALSE,
  cluster = NULL)

\method{partialCorrelation}{mmm}(object, var, fisher.transform = FALSE,
  cluster = NULL)
}
\arguments{
\item{object}{an \code{lm} object.}

\item{var}{[character] the covariate with which the partial correlation should be computed.}

\item{fisher.transform}{[logical] should the p-value/confidence intervals be computed using Fisher's Z transform.
Otherwise a Student's t-distribution is used}
}
\description{
Compute the partial correlation between the outcome and a covariate in a linear model.
}
\examples{
n <- 1e2
set.seed(10)
df <- data.frame(Y = rnorm(n),
                 X = rnorm(n),
                 K = as.character(rbinom(n, size = 3, prob = 0.5)))

## 1 covariate
e1.lm <- lm(Y~X, data = df)
partialCorrelation(e1.lm, var = "X")
cor.test(df$Y,df$X) ## same p-value, different CI

## 2 covariates
e2.lm <- lm(Y~X+K, data = df)
partialCorrelation(e2.lm, var = "X")

}
