% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_plotInteraction.R
\name{partialModel}
\alias{partialModel}
\title{Create a partial model relative to two variables}
\usage{
partialModel(object, var1, var2, method.coef = NULL, field.formula = NULL,
  field.data = NULL)
}
\arguments{
\item{object}{the model to be reduced}

\item{var1}{a covariate}

\item{var2}{another covariate}

\item{method.coef}{the method to extract the coefficient from the model}

\item{field.formula}{the name of the formula field in the call}

\item{field.data}{the name of the data field in the call}
}
\value{
the reduced model
}
\description{
Create a partial model relative to two variables
}
\examples{
library(lava)
library(nlme)

## lm
m <- lvm(Y~X1+X2+X3+X4+X5)
categorical(m, K=3) <- ~X1
categorical(m, K=2, labels = letters[1:2]) <- ~X2
d <- sim(m, 1e2)

lm.full <- lm(Y~X1*X2+X3+X4+X5, data = d)
coef(lm.full)
partialModel(lm.full, var1 = "X4", var2 = "X5")
partialModel(lm.full, var1 = "X1", var2 = "X5")
partialModel(lm.full, var1 = "X2", var2 = "X3")
partialModel(lm.full, var1 = "X1", var2 = "X2")

gls.full <- gls(Y~X1*X2+X3+X4+X5, data = d,
                weights = varIdent(form = ~ 1|X2))
coef(gls.full)
partialModel(gls.full, var1 = "X4", var2 = "X5")
partialModel(gls.full, var1 = "X1", var2 = "X5")
partialModel(gls.full, var1 = "X2", var2 = "X3")
partialModel(gls.full, var1 = "X1", var2 = "X2")

lme.full <- lme(Y~X1*X3+X4+X5, data = d, random = ~ 1|X2)
fixef(lme.full)
partialModel(lme.full, var1 = "X4", var2 = "X5")
partialModel(lme.full, var1 = "X1", var2 = "X5")
partialModel(lme.full, var1 = "X2", var2 = "X3")
partialModel(lme.full, var1 = "X1", var2 = "X2")

}

