% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.detectOutlier.R
\name{plot.detectOutlier}
\alias{plot.detectOutlier}
\title{plot function for detectOutlier objects}
\usage{
\method{plot}{detectOutlier}(x, ...)
}
\arguments{
\item{x}{an object of class detectOutlier.}

\item{...}{not used.}
}
\description{
plot function for detectOutlier objects.
}
