% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_plotInteraction.R
\name{plotInteraction}
\alias{plotInteraction}
\title{Display an first order interaction for categorical variables}
\usage{
plotInteraction(object, var1, var2, method.predict = NULL,
  field.data = NULL, alpha = 0.05, n.points = 1000, plot = TRUE,
  alpha.ggplot = 0.25, ...)
}
\arguments{
\item{object}{a linear model}

\item{var1}{a continuous or categorical covariate}

\item{var2}{a categorical covariate}

\item{method.predict}{a method to compute the predicted value from the model (fit) with their standard error (se.fit).}

\item{field.data}{the name of the data field in the call}

\item{alpha}{the alpha risk for the confidence interval}

\item{n.points}{the number of points use for the predictions along the x axis. Only relevant if var1 is numerical.}

\item{plot}{should the plot be displayed.}

\item{alpha.ggplot}{the transparency level for the display of the confidence intervals.}

\item{...}{additional argument to be passed to \code{partialModel}}
}
\description{
Display an first order interaction for categorical variables
}
\examples{
library(lava)
library(nlme)

## simulate interaction
m <- lvm(Y ~ X1 + gender + group + Interaction)
distribution(m, ~gender) <- binomial.lvm()
distribution(m, ~group) <- binomial.lvm()
constrain(m, Interaction ~ gender + group) <- function(x){x[,1]*x[,2]}
d <- sim(m, 1e2)
d$gender <- factor(d$gender, labels = letters[1:2])
d$group <- factor(d$group)

## linear model
lmfit <- lm(Y ~ X1 + gender*group, data = d)

plotInteraction(lmfit, var1 = "gender", var2 = "group")
plotInteraction(lmfit, var1 = "X1", var2 = "group")

## gls model
glsfit <- gls(Y ~ X1 + gender*group, data = d,
             weight = varIdent(form = ~1|group))

plotInteraction(glsfit, var1 = "gender", var2 = "group")
plotInteraction(glsfit, var1 = "X1", var2 = "group")

## mixed model
lmefit <- lme(Y ~ X1 + gender*group, data = d,
              random = ~1|group)

plotInteraction(lmefit, var1 = "gender", var2 = "group")
res <- plotInteraction(lmefit, var1 = "X1", var2 = "group")
res$model

}

