% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveTable.R
\name{print.descriptiveTable}
\alias{print.descriptiveTable}
\title{Print Function for the Descriptive Table}
\usage{
\method{print}{descriptiveTable}(object, print = TRUE,
  digit.frequency = 2, digit.center = 2, digit.spread = 2,
  format.date = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{object}{output of \code{descriptiveTable}.}

\item{print}{[logical] should the descriptive table be printed?}

\item{digit.frequency}{[integer, >=0] number of digit when printing frequencies.}

\item{digit.center}{[integer, >=0] number of digit when printing center parameters.}

\item{digit.spread}{[integer, >=0] number of digit when printing spread parameters.}

\item{format.date}{[character] the format used to output dates.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Print function for the descriptive table.
}
