% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Bootstrap.R
\name{print.glsboot}
\alias{print.glsboot}
\title{Display the result fo the bootstrap computation}
\usage{
\method{print}{glsboot}(x, seq_length.out, ...)
}
\arguments{
\item{x}{object obtained with the function \code{bootGLS}}

\item{seq_length.out}{the number of subsamples on which the results should be computed}

\item{...}{not used}
}
\description{
Display the result fo the bootstrap computation
}
\details{
Should have an attribute call that includes a data attribute (i.e. object$call$data must exist).
}
