% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot2.R
\name{qqplot2}
\alias{qqplot2}
\alias{qqplot2.lvmfit}
\alias{qqplot2.list}
\alias{qqplot2.multigroupfit}
\title{QQplot for the Residuals of a LVM Object}
\usage{
qqplot2(object, ...)

\method{qqplot2}{lvmfit}(object, variables = NULL, residuals = NULL,
  plot = TRUE, mfrow = NULL, mar = c(2, 2, 2, 2),
  qq.type = "qqtest", name.model = "", centralPercents = 0.95, ...)

\method{qqplot2}{list}(object, plot = TRUE, mfrow = NULL, mar = c(2,
  2, 2, 2), qq.type = "qqtest", name.model = "",
  centralPercents = 0.95, ...)

\method{qqplot2}{multigroupfit}(object, residuals = NULL,
  name.model = NULL, plot = TRUE, ...)
}
\arguments{
\item{object}{a lvm model.}

\item{...}{additional arguments to be passed to qqtest.}

\item{variables}{the variable for which the residuals should be displayed.}

\item{residuals}{[matrix] the residuals relative to each endogenous variable.}

\item{plot}{[logical] should the graphic be displayed?}

\item{mfrow}{how to divide the window. See \code{par}.}

\item{mar}{[numeric vector] the number of lines of margin
to be specified on the four sides of the plot (bottom, left, top, right).}

\item{qq.type}{the function used to display the qqplot. Can be qqtest or qqnorm.}

\item{name.model}{[character vector] character string to be displayed before the variable name in the title of the plot.
If \code{NULL} and \code{object} is a list, the names of the list will be used to define the title of the plots.}

\item{centralPercents}{argument passed to \code{qqtest}. See the help of \code{qqtest}.}
}
\value{
a data frame/cvlvm object containing the convergence status (by default 0 indicates successful convergence, see ?optim), the value of the log-likelihood and the estimated parameters (in columns) for each initialization (in rows)
}
\description{
QQplot for the residuals of a lvm object.
}
\details{
Simulation is based on a multivariate truncated normal law (even though it is not satifying for the variance components)
}
\examples{
#### lvm object ####
library(lava)
m <- lvm(list(y~v1+v2+v3+v4,c(v1,v2,v3,v4)~x))
latent(m) <- ~ x

set.seed(10)
dd <- sim(m,100) ## Simulate 100 observations from model
e <- estimate(m, dd) ## Estimate parameters

qqplot2(e)

#### multigroup object ####
e2 <- estimate(list(m1=m,m2=m), split(dd, dd$v1>0)) ## Estimate parameters

qqplot2(e2)

}
