% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleOutlier.R
\name{scaleOutlier}
\alias{scaleOutlier}
\title{Robust scaling}
\usage{
scaleOutlier(x, center = "median", scale = "mad", method,
  na.rm = FALSE, noScaleIf0 = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{center}{the method used to assess the "average" value}

\item{scale}{the method used to assess the dispersion of the data}

\item{method}{a method taking the dataset as its first argument and returning first the center and then the scale value. Disregarded if arguments center and scale are not null.}

\item{na.rm}{should missing values be omitted}

\item{noScaleIf0}{If \code{TRUE} the variable will not be scaled if its dispersion is null.}

\item{...}{additional arguments passed to method}
}
\value{
the scaled vector
}
\description{
Robust scaling of a numeric vector
}
\examples{
n <- 1e3
scaleOutlier(rnorm(n))

}
