% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2Dimage.R
\name{sim2Dimage}
\alias{sim2Dimage}
\title{Simulate a 2D Image}
\usage{
sim2Dimage(n, coords = NULL, x.sites, y.sites, mu, vgm, name.X = "X")
}
\arguments{
\item{n}{[integer, >0] number of images to be simulated}

\item{coords}{[2 column data.frame] coordinates of the element in the image.}

\item{x.sites}{[integer, >0] number of x-coordinates. Ignored when \code{coords} is specified.}

\item{y.sites}{[integer, >0] number of y-coordinates. Ignored when \code{coords} is specified.}

\item{mu}{[numeric vector] Expected value at each site.}

\item{vgm}{[variogramModel] Variogram used to generate the background noise. Output of the \code{vgm} function.}

\item{name.X}{[character] name of the column for the design matrix.}
}
\value{
A list with two elements \itemize{
\item coords: the coordinates of the sites
\item X: the design matrix containing the signal of each site for each replicate.
}
}
\description{
Simulate a 2D image with a given mean and autocorrelation structure.
}
\examples{

if(require(sp) & require(gstat)){
out0 <- sim2Dimage(3, x.sites = 50, y.sites = 50,
                  mu = 0,
                 vgm = vgm(psill = 1, range = 2, model='Exp'))
dim(out0$X)
spdf0 <- cbind(out0$coords,out0$X[1,])
gridded(spdf0) <- ~x+y
plot(spdf0)

coords <- expand.grid(x = 1:50,
                      y = 1:50)
coords$mu <- 0
coords$mu[coords$x<=5 | coords$y<=5] <- 2

out1 <- sim2Dimage(3, coords = coords[,c("x","y")],
                  mu = coords$mu,
                 vgm = vgm(psill = 1, range = 2, model='Exp'))
spdf1 <- cbind(out1$coords,out1$X[1,])
gridded(spdf1) <- ~x+y
plot(spdf1)

}
}
