% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootReg.R
\name{summary.bootReg}
\alias{summary.bootReg}
\title{Display the result fo the bootstrap computation}
\usage{
\method{summary}{bootReg}(object, p.value = TRUE, conf = 0.95,
  type = NULL, index = NULL, print = TRUE, n.subset = NULL, ...)
}
\arguments{
\item{object}{object obtained with the function \code{bootReg}}

\item{p.value}{should the p.value be computed? Can be time consuming.}

\item{conf}{the confidence level of the confidence intervals.}

\item{type}{the method used to compute the confidence intervals.
See the documentation of \code{boot::boot.ci}.}

\item{index}{for which parameters confidence intervals/p.values should be computed? Default is for all parameters.}

\item{print}{should the summary be displayed in the terminal?}

\item{n.subset}{if not \code{NULL} display in addition intervals/p.values for subset of the bootstrap samples.}

\item{...}{not used}
}
\description{
Display the result fo the bootstrap computation
}
\details{
This function calls the \code{boot.ci} from the boot package
to compute the confidence intervals.

P.value are computed by finding the confidence level
for which one bound of the confidence interval touches 0.

When the number of bootstrap sample is too low, the function (in fact \code{boot.ci}) will return warnings like
"extreme order statistics used as endpoints".
}
