% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{summary.butilsAccount}
\alias{summary.butilsAccount}
\alias{print.butilsAccount}
\title{Summarizing an account.}
\usage{
\method{print}{butilsAccount}(x, ...)

\method{summary}{butilsAccount}(
  object,
  print = TRUE,
  print.title = TRUE,
  detail = 0:2,
  keep.cols = c("paid", "date", "type", "total.cost", "participant.cost"),
  digit = 1,
  start = NULL,
  stop = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class butilsAccount.}

\item{...}{ignored}

\item{object}{an object of class butilsAccount.}

\item{print}{[logical] should the summary be printed in the console.}

\item{print.title}{[logical] should the title be printed in the console.}

\item{detail}{[logical] should the balance per individual (1) and by activity (2) be displayed.}

\item{keep.cols}{[character vector] the columns to be displayed in the detail.}

\item{digit}{[integer] the number of decimal to be displayed.}

\item{start}{[date, d-m-Y] define the start of the period where the balance should be computed.}

\item{stop}{[date, d-m-Y] define the end of the period where the balance should be computed.}
}
\description{
Summarizing an account.
}
