% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{update_package_files}
\alias{update_package_files}
\title{Automatically update package files}
\usage{
update_package_files(quiet = TRUE)
}
\arguments{
\item{quiet}{Should all output be suppressed?}
}
\description{
Automatically update package files
}
