% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_PreparePackage.R
\name{writeCollate_package}
\alias{writeCollate_package}
\title{write the collate field}
\usage{
writeCollate_package(package, path.package = NULL, space = "    ",
  trace = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path.package}{the position of the directory containing the package}
}

