% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{addNickName<-.butilsAccount}
\alias{addNickName<-.butilsAccount}
\alias{addActivity}
\alias{addActivity<-}
\alias{addActivity<-.butilsAccount}
\title{Add a new entry to the account.}
\usage{
\method{addNickName}{butilsAccount}(object) <- value

addActivity(object, ...) <- value

\method{addActivity}{butilsAccount}(object, involved, type = as.character(NA),
  date = as.Date(NA), ...) <- value
}
\arguments{
\item{object}{the account.}

\item{value}{a named vector describing name paid what.}

\item{involved}{who was involved in the activity?}

\item{type}{a character string describing the activity.}

\item{date}{the date at which the activity happen.}
}
\description{
Add a new entry to the account.
}
\examples{
myAcc <- createAccount()
addNickName(myAcc) <- list("Brice Ozenne" = "Brice",
                           "Sebastian Holst" = c("Seb","Sebastian"))
addActivity(myAcc, involved = c("Brice","Sebastian"), type = "Court") <- c("Sebastian" = 95)
myAcc
}
