% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{addNickName}
\alias{addNickName}
\alias{addNickName<-}
\title{Add nick name to the account.}
\usage{
addNickName(object) <- value
}
\arguments{
\item{object}{the account.}

\item{value}{a named vector.}
}
\description{
Add nick name to the account.
}
\examples{
myAcc <- createAccount()
addNickName(myAcc) <- list("Brice Ozenne" = "Brice",
                           "Sebastian Holst" = c("Seb","Sebastian"))
myAcc
}
