% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.R
\name{autoplot.breakpoint}
\alias{autoplot.breakpoint}
\title{Display Regression Line and Observations}
\usage{
\method{autoplot}{breakpoint}(object, pattern = NULL, plot = TRUE,
  combine.plot = TRUE, nrow = NULL, ncol = NULL, title = NULL,
  text.size = 10, add.cv.title = TRUE, add.bic.title = FALSE, ...)
}
\arguments{
\item{object}{output of \code{breakpoint}}

\item{pattern}{[vector of character] the number and type of breakpoints to be display.}

\item{plot}{[logical] should the plot be displayed in a window?}

\item{combine.plot}{[logical] should the plots for the different patterns be combined into one.}

\item{nrow}{[integer, >0] number of rows used when combining the plots.}

\item{ncol}{[integer, >0] number of columns used when combining the plots.}

\item{title}{[character] the title of the combined plot.}

\item{text.size}{[numeric, >0] the size of the text in the plot.}

\item{add.cv.title}{[logical] should the convergence status of the estimation algorithm
be displayed in the title of the plot.}

\item{add.bic.title}{[logical] should the bic of the model
be displayed in the title of the plot.}

\item{...}{not used. For compatibility with the generic function.}
}
\description{
Display regression line and observations.
}
