% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Bootstrap.R
\name{bootGLS}
\alias{bootGLS}
\title{Perform bootstrap computation on an object}
\usage{
bootGLS(object, n.boot, fctCoef, n.cpus = 1, IDvar = NULL,
  GROUPvar = NULL, load.library = "nlme", seed = 10)
}
\arguments{
\item{object}{the fitted model.}

\item{n.boot}{the number of replications. Should be a large number.}

\item{fctCoef}{the function used to extract the statistics of interest from the model.}

\item{n.cpus}{the number of cpu to use.}

\item{IDvar}{the variable in the dataset indicating the grouping level of the data.}

\item{GROUPvar}{the group variable use to resample under H0. Otherwise resampling is done under H1.}

\item{load.library}{additional library to load}

\item{seed}{set the random number generator}
}
\description{
Perform bootstrap computation on an object under H0 or H1. Handle one grouping variable.
}
\details{
Should have an attribute call that includes a data attribute (i.e. object$call$data must exist).

Bootstrap under H1: randomly select observations (or individuals according to argument IDvar) to form a new dataset.
If the same individual appear several times, a different group value is given for each apparition.

Bootstrap under H0: randomly permute the group label (according to argument GROUPvar) at the observation level (or individuals according to argument IDvar).
}
\examples{
#### lm ####
set.seed(10)
n <- 100
n.boot <- 1e4
Y1 <- rnorm(n, mean = 0)
Y2 <- rnorm(n, mean = 0.3)
df <- rbind(data.frame(Y=Y1,G=1),
           data.frame(Y=Y2,G=2)
           )
m.lm <- lm(Y ~ G, data = df)
summary(m.lm)

\dontshow{
res <- bootGLS(m.lm, n.boot = 1e2)
resG <- bootGLS(m.lm, n.boot = 1e2, GROUPvar = "G")

res
print(res, seq_length.out = 10)
resG 
}


\dontrun{ 
res <- bootGLS(m.lm, n.boot = 1e4, n.cpus = 4)
resG <- bootGLS(m.lm, n.boot = 1e4, GROUPvar = "G", n.cpus = 4)

res
print(res, seq_length.out = 10)
resG
}

#### gls ####
if(require("nlme")){
data(Ovary)
fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
             correlation = corAR1(form = ~ 1 | Mare))
summary(fm1)

\dontshow{
bootGLS(fm1, n.boot = 1e2, IDvar = "Mare")
}
\dontrun{
res <- bootGLS(fm1, n.boot = 1e4, n.cpus = 4, IDvar = "Mare")
apply(res$all.boot, 2, mean)
}

}

}

