% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_preparePackage.R
\name{buildPackage}
\alias{buildPackage}
\title{Integrative function for building packages}
\usage{
buildPackage(package, version = NULL, path.package = NULL,
  compileAttributes = TRUE, clearDir = TRUE, roxygenise = TRUE,
  build = TRUE, options.build = NULL, untar = TRUE, check = FALSE,
  options.check = NULL, install = TRUE, options.install = "--build",
  trace = 2, clearAfter = TRUE)
}

