% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialResiduals.R
\name{calcPartialResiduals}
\alias{calcPartialResiduals}
\title{Deprecated Function Computing Partial Residuals}
\usage{
calcPartialResiduals(...)
}
\arguments{
\item{...}{not used.}
}
\description{
Deprecated function computing partial residuals,
use \code{\link{partialResiduals}} instead.
}
