% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPartialResiduals.R
\name{calcPartialResiduals}
\alias{calcPartialResiduals}
\title{Compute partial residuals for linear and linear mixed models.}
\usage{
calcPartialResiduals(model, var, conditional = FALSE,
  interval = "confidence", level = 0.95, npoints = 100,
  quantile.range = c(0, 1), FUN.df, ...)
}
\arguments{
\item{model}{Model object (e.g. \code{lm})}

\item{var}{Variable relative to which the partial residual will be displayed.}

\item{conditional}{are the predictions conditional to the random effect? (if any)}

\item{interval}{Type of interval calculation ("confidence" or "prediction").}

\item{level}{Level of confidence limits (default 95\%)}

\item{npoints}{Length of the vector of unique values relative to each continuous variable.}

\item{quantile.range}{the quantiles (for the continous covariates) between which the fitted values will be computed.}

\item{FUN.df}{Optional function returning the residual degree of freedoms.}

\item{\dots}{additional arguments to lower level functions}
}
\value{
list with following members:
\item{data}{Original dataset with an additional column containing the partial residuals}
\item{partialFit}{Fitted values according to the range of values of var}
\item{var}{variable(s) for which the partial residuals have been computed}
\item{level}{Level of confidence}
\item{name.Y}{name of the response variable}
}
\description{
Compute partial residuals for linear and linear mixed models.
}
\details{
IMPORTANT: When using mixed models, the confidence and prediction intervals ignore the uncertainty of the covariance parameters.
}
\examples{
library(lava)
set.seed(10)
m.lvm <- lvm(Y~X1+X2+Id+Id2)
categorical(m.lvm, K = 5) <- ~Id+Id2
d <- lava::sim(n = 1e2, m.lvm)
 
m <- lm(Y~X1+X2, data = d)
pres1 <- calcPartialResiduals(m, var = "X1")
pres2 <- calcPartialResiduals(m, var = c("X1","X2"))


}
\seealso{
\code{plot.partialResiduals} for a graphical display and more examples.'
}
\author{
Brice Ozenne
}
\keyword{regression}
