% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{create_package_files}
\alias{create_package_files}
\title{Automatically create package files}
\usage{
create_package_files(quiet)
}
\arguments{
\item{quiet}{If \code{TRUE}, all output is suppressed.}
}
\description{
Automatically create package files
}
