% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalInParentEnv.R
\name{evalInParentEnv}
\alias{evalInParentEnv}
\title{Find object in the parent environments}
\usage{
evalInParentEnv(name)
}
\arguments{
\item{name}{character string containing the name of the object to get.}
}
\description{
Internal function
}
