% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Batch.R
\name{launchBatch}
\alias{launchBatch}
\title{Launching batch}
\usage{
launchBatch(path = ".", file, dirBatch = NULL, txt.options,
  operator = "start R CMD BATCH", add.Rout = TRUE, add.lis = TRUE,
  add.options = "", rm.newfile = FALSE)
}

