% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPower_ttest.R
\name{logPower_ttest}
\alias{logPower_ttest}
\title{Power Calculation For t-tests on Log-Transformed Data}
\usage{
logPower_ttest(
  mu.original,
  sigma2.original,
  gamma,
  rho = NULL,
  n = NULL,
  ratio.n = 1,
  sig.level = 0.05,
  power = 0.8,
  type = "two.sample",
  equivalence = FALSE,
  method.meanvar = "lognorm",
  method.cor = "uniroot",
  n.large = 1e+05,
  trace = TRUE,
  ncpus = NULL
)
}
\arguments{
\item{mu.original}{[numeric] expected value of the outcome on the original scale for the control group or baseline.}

\item{sigma2.original}{[numeric] variance of the outcome on the original scale. Assumed to be the same over groups or over time.}

\item{gamma}{[numeric] relative change for the expected value in the treatment group or follow-up value.}

\item{rho}{[numeric 0-1] correlation coefficient between the paired values on the original scale. Only used when \code{type="paired"} or \code{type="equivalence"}.}

\item{n}{[integer >1] Sample size.}

\item{ratio.n}{[numeric >1] ratio between the sample size in larger and the smaller group. By default 1.
Note that the larger group is the reference group (with expectation \code{mu.original}) and the smaller group the active group (with expectation \code{gamma*mu.original}).}

\item{sig.level}{[numeric 0-1] type 1 error}

\item{power}{[numeric 0-1] statistical power (i.e. complement to 1 of the type 2 error)}

\item{type}{[character] type of study:
difference in means between two independent groups (\code{type="two.sample"}),
difference in means between paired measurements (\code{type="paired"}),
equivalence in means between paired measurements (\code{type="equivalence"}),}

\item{method.meanvar}{[character] method used to identify the expected mean and variance on the log scale.
Either assumes that the outcome is log-normally distributed on the original scale (\code{method.meanvar = "lognorm"}),
or that it is normally distributed on the log-scale (\code{method.meanvar = "lognorm2"}).}

\item{method.cor}{[character] method used to identify the correlation on the log scale: one of \code{"uniroot"}, \code{"optim"}, or \code{"taylor"}.}

\item{n.large}{[integer, >0] sample size used to indentify the correlation coefficient or assess the error made when identifying the parameters. Should be large.}

\item{trace}{[logical] Should a progress bar be displayed when estimating the power for various combinaisons of parameters?}

\item{ncpus}{[integer, >0] Number of cores to be used, i.e., how many processes can be run simultaneously.}
}
\description{
Power calculation for t-tests on log-transformed data.
}
\references{
Shein-Chung Chow , Jun Shao & Hansheng Wang (2002). A note on sample size calculation for mean comparisons based on noncentral t-statistics, Journal of Biopharmaceutical Statistics, 12:4, 441-456, DOI: 10.1081/BIP-120016229
}
