% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_detectOutliers.R
\name{numOutlier}
\alias{numOutlier}
\title{Identify outlier}
\usage{
numOutlier(x, threshold = NULL, type = "auto", na.rm = FALSE)
}
\examples{
detectOutlier(rnorm(1e3))
}
\references{
idea from http://www.r-bloggers.com/finding-outliers-in-numerical-data/
}

