% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_DetectOutliers.R
\name{numOutlier}
\alias{numOutlier}
\title{Identify outlier}
\usage{
numOutlier(x, type = "auto", th.gaussian = 3, th.hampel = 3,
  th.boxplot = 1.5, na.rm = FALSE)
}
\examples{
numOutlier(rnorm(1e3))
}
\references{
idea from http://www.r-bloggers.com/finding-outliers-in-numerical-data/
}

