% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Source.R
\name{package.source}
\alias{package.source}
\title{Source a package directory}
\usage{
package.source(name, path = dir.gitHub(), Rcode = TRUE, Ccode = FALSE)
}
\arguments{
\item{name}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{Rcode}{should all the .R be sourced}

\item{Ccode}{should all the .cpp file be source (using Rcpp::sourceCpp)}
}
\description{
Source all the R and Cpp file contain in a package
}
\seealso{
\code{\link{dir.gitHub}}
}

