% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.partialResiduals.R
\name{plot.partialResiduals}
\alias{plot.partialResiduals}
\title{Display an first order interaction for categorical variables}
\usage{
\method{plot}{partialResiduals}(x, alpha.ggplot = 0.25, ...)
}
\arguments{
\item{x}{a linear model}

\item{alpha.ggplot}{transparency parameter for the points or the bands}

\item{...}{additional argument to be passed to \code{partialModel}}
}
\description{
Display an first order interaction for categorical variables
}
\examples{
library(lava)
set.seed(10)
m.lvm <- lvm(Y~2*Age+4*gender+gene+time)
categorical(m.lvm, labels = c("M","F")) <- ~gender
categorical(m.lvm, K = 10) <- ~gene
d <- lava::sim(n = 1e2, m.lvm)

## linear model
m <- lm(Y~Age+gender, data = d)
pres1 <- calcPartialResiduals(m, var = "Age")
plot(pres1)
pres2 <- calcPartialResiduals(m, var = c("Age","gender"))
plot(pres2)
pres3 <- calcPartialResiduals(m, var = c("Age","gender"), interval = "prediction")
plot(pres3)
pres4 <- calcPartialResiduals(m, var = "gender")
plot(pres4)

## linear mixed model
if(require(lme4) && require(merTools) && require(AICcmodavg)){

mm <- lmer(Y~Age+gender+(1|gene), data = d)
pres1 <- calcPartialResiduals(mm, var = "Age")
plot(pres1)
pres2 <- calcPartialResiduals(mm, var = c("Age","gender"))
plot(pres2)

# using external function
pres3 <- calcPartialResiduals(mm, var = c("Age","gender"), FUN.predict = predict_merTools)
plot(pres3) 
pres4 <- calcPartialResiduals(mm, var = c("Age","gender"), FUN.predict = predict_AICcmodavg)
plot(pres4)

}
}
