% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveTable.R
\name{print.descriptiveTable}
\alias{print.descriptiveTable}
\title{Print Function for the Descriptive Table}
\usage{
\method{print}{descriptiveTable}(
  x,
  print = TRUE,
  print.minmax = TRUE,
  print.spread = TRUE,
  digit.frequency = 2,
  digit.center = 2,
  digit.spread = 2,
  digit.p = 2,
  format.date = "\%Y-\%m-\%d",
  ...
)
}
\arguments{
\item{x}{output of \code{descriptiveTable}.}

\item{print}{[logical] should the descriptive table be printed?}

\item{print.minmax}{[logical] should minimum and maximum values be displayed?}

\item{print.spread}{[logical] should the spread of the values be displayed?}

\item{digit.frequency}{[integer, >=0] number of digit when printing frequencies.}

\item{digit.center}{[integer, >=0] number of digit when printing center parameters.}

\item{digit.spread}{[integer, >=0] number of digit when printing spread parameters.}

\item{digit.p}{[integer, >=0] number of digit when printing p-values.}

\item{format.date}{[character] the format used to output dates.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Print function for the descriptive table.
}
