% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Profile.R
\name{profileCode}
\alias{profileCode}
\title{Profile a function}
\usage{
profileCode(FUN, filename = NULL, type = "Rprof", plot = FALSE, ...)
}
\arguments{
\item{FUN}{the function to be profiled. Must have no argument}

\item{filename}{the name of the file where to save the profiling. I}

\item{type}{the function used for profiling: can be \code{"Rprof"} or \code{"profvis"}.}

\item{plot}{should the result of the profiling be graphically displayed.}

\item{...}{additional arguments to be passed to the profiling function}
}
\value{
invisible output of the profiling method.
}
\description{
This function is a convenient to profile the execution of a function
}
\details{
If \code{filename} is \code{NULL} a temporary file is created and is removed at the end of the execution of the function.
}
\examples{

\dontrun{
library(lava)
m <- lvm(Y ~ X + G)
d <- sim(m, 1e2)
FUN <- function(){
estimate(m,d)
}
profileCode(FUN, type = "Rprof")
profileCode(FUN, type = "profvis")
profileCode(FUN, type = "profvis", plot = TRUE)
}

}
\keyword{function}
\keyword{profile}

