% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_DetectOutliers.R
\name{scanOutlier}
\alias{scanOutlier}
\title{Search potential outliers in a dataset}
\usage{
scanOutlier(data, id, method.numeric = "numOutlier",
  method.factor = "factorOutlier", method.id = "checkId", args.num = NULL,
  args.factor = NULL)
}
\examples{

}

