% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilite.R
\name{summary.butilsAccount}
\alias{summary.butilsAccount}
\alias{print.butilsAccount}
\alias{summary.butilsAccount}
\title{Summarizing an account.}
\usage{
\method{print}{butilsAccount}(x, ...)

\method{summary}{butilsAccount}(object, print = TRUE, detail = 0:2,
  keep.cols = c("paid", "date", "type", "total.cost", "participant.cost"),
  digit = 1, ...)
}
\arguments{
\item{x}{an object of class butilsAccount.}

\item{...}{ignored}

\item{object}{an object of class butilsAccount.}

\item{print}{should the summary be printed in the console.}

\item{detail}{should the balance per individual (1) and by activity (2) be displayed.}

\item{keep.cols}{the columns to be displayed in the detail.}
}
\description{
Summarizing an account.
}
