% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_magic.R
\name{validate_package_files}
\alias{validate_package_files}
\title{Automatically update package files}
\usage{
validate_package_files(quiet, error_on)
}
\arguments{
\item{quiet}{If \code{TRUE}, all output is suppressed.}

\item{error_on}{Whether to throw an error on \code{R CMD check} failures.
Note that the check is always completed (unless a timeout happens),
and the error is only thrown after completion. If \code{"never"}, then
no errors are thrown. If \code{"error"}, then only \code{ERROR} failures
generate errors. If \code{"warning"}, then \code{WARNING} failures generate
errors as well. If \code{"note"}, then any check failure generated an
error.}
}
\description{
This function is read-only.
}
