% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Valid.R
\name{validFCTs}
\alias{validCharacter}
\alias{validClass}
\alias{validDimension}
\alias{validFCTs}
\alias{validInteger}
\alias{validLogical}
\alias{validNames}
\alias{validNumeric}
\alias{validPath}
\title{check the validity of the arguments given by the user}
\usage{
validCharacter(value, name = as.character(substitute(value)), validLength,
  validValues = "character", refuse.NULL = TRUE,
  refuse.duplicates = FALSE, method = NULL, addPP = TRUE)

validClass(value, name = as.character(substitute(value)), validClass,
  superClasses = TRUE, method = NULL, addPP = TRUE)

validDimension(value1, value2 = NULL,
  name1 = as.character(substitute(value1)),
  name2 = as.character(substitute(value2)), validDimension = NULL,
  type = c("NROW", "NCOL"), method = NULL, addPP = TRUE)

validInteger(value, name = as.character(substitute(value)), validLength,
  validValues = NULL, min = NULL, max = NULL, refuse.NA = TRUE,
  refuse.NULL = TRUE, refuse.duplicates = FALSE, method = NULL,
  addPP = TRUE)

validLogical(value, name = as.character(substitute(value)), validLength,
  refuse.NULL = TRUE, refuse.NA = TRUE, method = NULL, addPP = TRUE)

validNames(value, name = as.character(substitute(value)),
  refuse.NULL = TRUE, validLength = NULL, validValues = NULL,
  requiredValues = NULL, forbiddenValues = NULL, method = NULL,
  addPP = TRUE)

validNumeric(value, name = as.character(substitute(value)), validLength,
  validValues = NULL, min = NULL, max = NULL, refuse.NA = TRUE,
  refuse.NULL = TRUE, refuse.duplicates = FALSE, method = NULL,
  addPP = TRUE)

validPath(value, name = as.character(substitute(value)), type,
  method = NULL, addPP = TRUE, extension = NULL, checkFsep = FALSE)
}

